refactor(agent_sre): rename dir + single editable package (lambda_src)
- Rename sre-incident-copilot/ -> agent_sre/
- All project Python now lives in the Lambda FUNCTION code as the lambda_src
  package (console-editable, ~10KB): agent_app.py (student edits), runtime,
  prompt, model, tools, knowledge_base, aws_docs, memory.
- Third-party libraries (strands/fastapi/uvicorn/mcp/awslabs) live ONLY in a
  Lambda layer (copilot-deps), published by package_and_deploy.sh.
- agent_app.py moved inside lambda_src; runtime imports it via 'from . import
  agent_app'; run.sh -> uvicorn lambda_src.runtime:app.
- Fix all path/name references (packager, faults/inject.sh, README, ingest UA).

Streaming retained (LWA). Deployed & verified: chat works (M1, no tools).
