+++ agent_sre/kb/setup_s3vectors.sh
